Introducing Metric Types
There are certain metrics that are INPUT only and not used for benchmarking (E.g. EBITDA) and some are calculated (Rule of 40) based on input and does not need direct input from user. In some other cases, it is required as input and directly used for benchmarking. To classify these, I am creating a metric types table and classifying each metric as input only, calculated only or benchmark.. INPUT and BENCHMARK TYPES should have sliders (ask for input from user) where as CALCULATED AND BENCHMARK will appear in reports.

To reduce the re-work, I am not changing the db_queries to filter automatically - rather adding attributes so filtering can happen on the front end. This is a potential performance optimization for future where we can separate the SQL for input and report generation (CQRS).